requirements.txt
ollama
faster-whisper
pyserial